Spec 31590: S3 folder import/export, incl. Files and experiments runs/data

+++ api/src/org/labkey/api/pipeline/LocalDirectory.java
@@ -129,8 +129,11 @@ public class LocalDirectory implements Serializable
                 try
                 {
                     File tempFile = new File(_localDirectoryFile, filename);
-                    Files.copy(path, tempFile.toPath(), StandardCopyOption.COPY_ATTRIBUTES);
-                    log.info("Created temp file because input is from cloud: " + FileUtil.pathToString(path));
+                    if (!Files.exists(tempFile.toPath()))
+                    {
+                        Files.copy(path, tempFile.toPath(), StandardCopyOption.COPY_ATTRIBUTES);
+                        log.info("Created temp file because input is from cloud: " + FileUtil.pathToString(path));
+                    }
                     return tempFile;
                 }
                 catch (IOException e)
